refactor: Path 생성 인터페이스로 분리
