+++ src/main/java/wooteco/subway/domain/PathStrategy.java
@@ -1,0 +1,8 @@
+package wooteco.subway.domain;
+
+import java.util.List;
+
+public interface PathStrategy {
+
+    Path findPath(List<Section> sections, Long source, Long target);
+}
